pyproject.toml
[project]
name = "hybrid-axon-seg"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "cellpose==4.0.9",
    "scikit-image",
    "scipy",
    "numpy<2",
    "pandas",
    "matplotlib",
    "pillow",
    "openpyxl",
    "torch",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pre-commit",
]

# ── Ruff ────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (unused imports, undefined names, …)
    "I",   # isort (import ordering)
    "UP",  # pyupgrade (modern Python syntax)
    "B",   # flake8-bugbear (likely bugs)
    "C4",  # flake8-comprehensions (unnecessary list/dict/set calls)
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long — handled by formatter
    "SIM108", # ternary operator — sometimes less readable
    "B007",   # loop variable not used in loop body — false positives with enumerate
]

[tool.ruff.lint.isort]
known-first-party = [
    "config", "utils", "preprocessing", "detection",
    "morphometrics", "qc", "visualization",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
